Return a proxy when traversing the help namespaces.
Fixes #4 by returning a new object, so we are thread-safe.

This new object is just a plain proxy, not a LocationProxy. Trying to
return a LocationProxy (with the __name__ taken from the traversal
name and the __parent__ set to the context) causes lots of 404 errors
from zope.app.apidoc (although no failures here). I suspect this is
because OnlineHelp is supposed to be an IContainmentRoot, which isn't
supposed to have a name.

Like LocationProxy, though, make the returned proxy not capable of
being pickled.